Refactor search bar to allow for different search engines
